pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amis_python"
version = "0.1.0"
description = "Python-side schema builder for amis JSON"
readme = "README.md"
requires-python = ">=3.8.0, <3.9" # 修改为 Python 3.8+
dependencies = [
    "django>=3.1.3",
    #    "django-extensions>=3.2.3",
    #    "django-ninja>=1.5.0",
    #    "django-ninja-extra>=0.30.6",
    #    "mypy>=1.14.1",
    "pydantic>=2.0,<2.11.0",
    #    "pytest>=8.3.5",
    "setuptools>=75.3.2",
    "typing-extensions>=4.6.1",
    'djangorestframework>=3.12.2'
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.workspace]

[tool.setuptools.package-data]
"*" = ["static/**/*"]
